+++ src/lib/filter/filter.test.ts
@@ -2,712 +2,1071 @@
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
 import { expect, it, describe } from "@jest/globals";
-import { kind, modelToGroupVersionKind } from "kubernetes-fluent-client";
+import { kind, KubernetesObject, modelToGroupVersionKind } from "kubernetes-fluent-client";
 import * as fc from "fast-check";
+import { clone } from "ramda";
 import { AdmissionRequestCreatePod, AdmissionRequestDeletePod } from "../../fixtures/loader";
 import {
-  shouldSkipRequest,
+  adjudicateCarriesIgnoredNamespace,
   adjudicateMisboundDeleteWithDeletionTimestamp,
+  adjudicateMisboundNamespace,
+  adjudicateMismatchedAnnotations,
   adjudicateMismatchedDeletionTimestamp,
   adjudicateMismatchedEvent,
-  adjudicateMismatchedNameRegex,
-  adjudicateMismatchedName,
   adjudicateMismatchedGroup,
-  adjudicateMismatchedVersion,
   adjudicateMismatchedKind,
+  adjudicateMismatchedLabels,
+  adjudicateMismatchedName,
+  adjudicateMismatchedNameRegex,
+  adjudicateMismatchedNamespace,
+  adjudicateMismatchedNamespaceRegex,
+  adjudicateMismatchedVersion,
+  adjudicateMissingCarriableNamespace,
   adjudicateUnbindableNamespaces,
   adjudicateUncarryableNamespace,
-  adjudicateMismatchedNamespace,
-  adjudicateMismatchedLabels,
-  adjudicateMismatchedAnnotations,
-  adjudicateMismatchedNamespaceRegex,
-  adjudicateCarriesIgnoredNamespace,
-  adjudicateMissingCarriableNamespace,
+  filterNoMatchReason,
+  shouldSkipRequest,
 } from "./filter";
 import { AdmissionRequest, Binding } from "../types";
 import { Event, Operation } from "../enums";
-import { clusterScopedBinding } from "../helpers.test";
-import { defaultAdmissionRequest, defaultBinding } from "./adjudicators/defaultTestObjects";
+import {
+  defaultAdmissionRequest,
+  defaultBinding,
+  defaultFilters,
+  defaultKubernetesObject,
+} from "./adjudicators/defaultTestObjects";
+
 const callback = () => undefined;
 
-const podKind = modelToGroupVersionKind(kind.Pod.name);
-
-describe("Fuzzing shouldSkipRequest", () => {
-  it("should handle random inputs without crashing", () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          event: fc.constantFrom("CREATE", "UPDATE", "DELETE", "ANY"),
-          kind: fc.record({
-            group: fc.string(),
-            version: fc.string(),
-            kind: fc.string(),
-          }),
-          filters: fc.record({
-            name: fc.string(),
-            namespaces: fc.array(fc.string()),
-            labels: fc.dictionary(fc.string(), fc.string()),
-            annotations: fc.dictionary(fc.string(), fc.string()),
-            deletionTimestamp: fc.boolean(),
-          }),
-        }),
-        fc.record({
-          operation: fc.string(),
-          uid: fc.string(),
-          name: fc.string(),
-          namespace: fc.string(),
-          kind: fc.record({
-            group: fc.string(),
-            version: fc.string(),
-            kind: fc.string(),
-          }),
-          object: fc.record({
-            metadata: fc.record({
-              deletionTimestamp: fc.option(fc.date()),
+describe("shouldSkipRequest", () => {
+  describe("Fuzzing shouldSkipRequest", () => {
+    it("should handle random inputs without crashing", () => {
+      fc.assert(
+        fc.property(
+          fc.record({
+            event: fc.constantFrom("CREATE", "UPDATE", "DELETE", "ANY"),
+            kind: fc.record({
+              group: fc.string(),
+              version: fc.string(),
+              kind: fc.string(),
+            }),
+            filters: fc.record({
+              name: fc.string(),
+              namespaces: fc.array(fc.string()),
+              labels: fc.dictionary(fc.string(), fc.string()),
+              annotations: fc.dictionary(fc.string(), fc.string()),
+              deletionTimestamp: fc.boolean(),
             }),
           }),
-        }),
-        fc.array(fc.string()),
-        (binding, req, capabilityNamespaces) => {
-          expect(() =>
-            shouldSkipRequest(binding as Binding, req as AdmissionRequest, capabilityNamespaces),
-          ).not.toThrow();
+          fc.record({
+            operation: fc.string(),
+            uid: fc.string(),
+            name: fc.string(),
+            namespace: fc.string(),
+            kind: fc.record({
+              group: fc.string(),
+              version: fc.string(),
+              kind: fc.string(),
+            }),
+            object: fc.record({
+              metadata: fc.record({
+                deletionTimestamp: fc.option(fc.date()),
+              }),
+            }),
+          }),
+          fc.array(fc.string()),
+          (binding, req, capabilityNamespaces) => {
+            expect(() =>
+              shouldSkipRequest(binding as Binding, req as AdmissionRequest, capabilityNamespaces),
+            ).not.toThrow();
+          },
+        ),
+        { numRuns: 100 },
+      );
+    });
+  });
+
+  describe("Property-Based Testing shouldSkipRequest", () => {
+    it("should only skip requests that do not match the binding criteria", () => {
+      fc.assert(
+        fc.property(
+          fc.record({
+            event: fc.constantFrom("CREATE", "UPDATE", "DELETE", "ANY"),
+            kind: fc.record({
+              group: fc.string(),
+              version: fc.string(),
+              kind: fc.string(),
+            }),
+            filters: fc.record({
+              name: fc.string(),
+              namespaces: fc.array(fc.string()),
+              labels: fc.dictionary(fc.string(), fc.string()),
+              annotations: fc.dictionary(fc.string(), fc.string()),
+              deletionTimestamp: fc.boolean(),
+            }),
+          }),
+          fc.record({
+            operation: fc.string(),
+            uid: fc.string(),
+            name: fc.string(),
+            namespace: fc.string(),
+            kind: fc.record({
+              group: fc.string(),
+              version: fc.string(),
+              kind: fc.string(),
+            }),
+            object: fc.record({
+              metadata: fc.record({
+                deletionTimestamp: fc.option(fc.date()),
+              }),
+            }),
+          }),
+          fc.array(fc.string()),
+          (binding, req, capabilityNamespaces) => {
+            const shouldSkip = shouldSkipRequest(binding as Binding, req as AdmissionRequest, capabilityNamespaces);
+            expect(typeof shouldSkip).toBe("string");
+          },
+        ),
+        { numRuns: 100 },
+      );
+    });
+  });
+
+  it("create: should reject when regex name does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "^default$",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines name regex '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("create: should not reject when regex name does match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "^cool",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("delete: should reject when regex name does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "^default$",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestDeletePod();
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines name regex '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("delete: should not reject when regex name does match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "^cool",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestDeletePod();
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("create: should not reject when regex namespace does match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [new RegExp("^helm").source],
+        regexName: "",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("create: should reject when regex namespace does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [new RegExp("^argo").source],
+        regexName: "",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines namespace regexes '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("delete: should reject when regex namespace does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [new RegExp("^argo").source],
+        regexName: "",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestDeletePod();
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines namespace regexes '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("delete: should not reject when regex namespace does match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [new RegExp("^helm").source],
+        regexName: "",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestDeletePod();
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("delete: should reject when name does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "bleh",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "^not-cool",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestDeletePod();
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines name '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("should reject when kind does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: {
+        group: "",
+        version: "v1",
+        kind: "Nope",
+      },
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines kind '.*' but Request declares '.*'./,
+    );
+  });
+
+  it("should reject when group does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: {
+        group: "Nope",
+        version: "v1",
+        kind: "Pod",
+      },
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines group '.*' but Request declares '.*'./,
+    );
+  });
+
+  it("should reject when version does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: {
+        group: "",
+        version: "Nope",
+        kind: "Pod",
+      },
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines version '.*' but Request declares '.*'./,
+    );
+  });
+
+  it("should allow when group, version, and kind match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should allow when kind match and others are empty", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: {
+        group: "",
+        version: "",
+        kind: "Pod",
+      },
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should reject when the capability namespace does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, ["bleh", "bleh2"])).toMatch(
+      /Ignoring Admission Callback: Object carries namespace '.*' but namespaces allowed by Capability are '.*'./,
+    );
+  });
+
+  it("should reject when namespace does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: ["bleh"],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines namespaces '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("should allow when namespace is match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: ["helm-releasename", "unicorn", "things"],
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should reject when label does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {
+          foo: "bar",
         },
-      ),
-      { numRuns: 100 },
+        annotations: {},
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines labels '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("should allow when label is match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        deletionTimestamp: false,
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "",
+        labels: {
+          foo: "bar",
+          test: "test1",
+        },
+        annotations: {},
+      },
+      callback,
+    };
+
+    const pod = AdmissionRequestCreatePod();
+    pod.object.metadata = pod.object.metadata || {};
+    pod.object.metadata.labels = {
+      foo: "bar",
+      test: "test1",
+      test2: "test2",
+    };
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should reject when annotation does not match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {
+          foo: "bar",
+        },
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+    const pod = AdmissionRequestCreatePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines annotations '.*' but Object carries '.*'./,
+    );
+  });
+
+  it("should allow when annotation is match", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        annotations: {
+          foo: "bar",
+          test: "test1",
+        },
+        deletionTimestamp: false,
+        regexNamespaces: [],
+        regexName: "",
+      },
+      callback,
+    };
+
+    const pod = AdmissionRequestCreatePod();
+    pod.object.metadata = pod.object.metadata || {};
+    pod.object.metadata.annotations = {
+      foo: "bar",
+      test: "test1",
+      test2: "test2",
+    };
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should use `oldObject` when the operation is `DELETE`", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.DELETE,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "",
+        deletionTimestamp: false,
+        labels: {
+          "test-op": "delete",
+        },
+        annotations: {},
+      },
+      callback,
+    };
+
+    const pod = AdmissionRequestDeletePod();
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should allow when deletionTimestamp is present on pod", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        regexNamespaces: [],
+        regexName: "",
+        annotations: {
+          foo: "bar",
+          test: "test1",
+        },
+        deletionTimestamp: true,
+      },
+      callback,
+    };
+
+    const pod = AdmissionRequestCreatePod();
+    pod.object.metadata = pod.object.metadata || {};
+    pod.object.metadata!.deletionTimestamp = new Date("2021-09-01T00:00:00Z");
+    pod.object.metadata.annotations = {
+      foo: "bar",
+      test: "test1",
+      test2: "test2",
+    };
+
+    expect(shouldSkipRequest(binding, pod, [])).toBe("");
+  });
+
+  it("should reject when deletionTimestamp is not present on pod", () => {
+    const binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: podKind,
+      filters: {
+        name: "",
+        namespaces: [],
+        labels: {},
+        regexNamespaces: [],
+        regexName: "",
+        annotations: {
+          foo: "bar",
+          test: "test1",
+        },
+        deletionTimestamp: true,
+      },
+      callback,
+    };
+
+    const pod = AdmissionRequestCreatePod();
+    pod.object.metadata = pod.object.metadata || {};
+    pod.object.metadata.annotations = {
+      foo: "bar",
+      test: "test1",
+      test2: "test2",
+    };
+
+    expect(shouldSkipRequest(binding, pod, [])).toMatch(
+      /Ignoring Admission Callback: Binding defines deletionTimestamp but Object does not carry it./,
     );
   });
 });
 
-describe("Property-Based Testing shouldSkipRequest", () => {
-  it("should only skip requests that do not match the binding criteria", () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          event: fc.constantFrom("CREATE", "UPDATE", "DELETE", "ANY"),
-          kind: fc.record({
-            group: fc.string(),
-            version: fc.string(),
-            kind: fc.string(),
-          }),
-          filters: fc.record({
-            name: fc.string(),
-            namespaces: fc.array(fc.string()),
-            labels: fc.dictionary(fc.string(), fc.string()),
-            annotations: fc.dictionary(fc.string(), fc.string()),
-            deletionTimestamp: fc.boolean(),
-          }),
-        }),
-        fc.record({
-          operation: fc.string(),
-          uid: fc.string(),
-          name: fc.string(),
-          namespace: fc.string(),
-          kind: fc.record({
-            group: fc.string(),
-            version: fc.string(),
-            kind: fc.string(),
-          }),
-          object: fc.record({
-            metadata: fc.record({
-              deletionTimestamp: fc.option(fc.date()),
-            }),
-          }),
-        }),
-        fc.array(fc.string()),
-        (binding, req, capabilityNamespaces) => {
-          const shouldSkip = shouldSkipRequest(binding as Binding, req as AdmissionRequest, capabilityNamespaces);
-          expect(typeof shouldSkip).toBe("string");
-        },
-      ),
-      { numRuns: 100 },
+export const podKind = modelToGroupVersionKind(kind.Pod.name);
+export const deploymentKind = modelToGroupVersionKind(kind.Deployment.name);
+export const clusterRoleKind = modelToGroupVersionKind(kind.ClusterRole.name);
+
+export const groupBinding: Binding = {
+  event: Event.CREATE,
+  filters: defaultFilters,
+  kind: deploymentKind,
+  model: kind.Deployment,
+};
+
+export const clusterScopedBinding: Binding = {
+  event: Event.DELETE,
+  filters: defaultFilters,
+  kind: clusterRoleKind,
+  model: kind.ClusterRole,
+};
+
+describe("filterNoMatchReason", () => {
+  it.each([
+    [{}],
+    [{ metadata: { namespace: "pepr-uds" } }],
+    [{ metadata: { namespace: "pepr-core" } }],
+    [{ metadata: { namespace: "uds-ns" } }],
+    [{ metadata: { namespace: "uds" } }],
+  ])(
+    "given %j, it returns regex namespace filter error for Pods whose namespace does not match the regex",
+    (obj: KubernetesObject) => {
+      const kubernetesObject: KubernetesObject = obj.metadata
+        ? {
+            ...defaultKubernetesObject,
+            metadata: { ...defaultKubernetesObject.metadata, namespace: obj.metadata.namespace },
+          }
+        : { ...defaultKubernetesObject, metadata: obj as unknown as undefined };
+      const binding: Binding = {
+        ...defaultBinding,
+        kind: { kind: "Pod", group: "some-group" },
+        filters: { ...defaultFilters, regexNamespaces: [new RegExp("(.*)-system").source] },
+      };
+
+      const capabilityNamespaces: string[] = [];
+      const expectedErrorMessage = `Ignoring Watch Callback: Binding defines namespace regexes '["(.*)-system"]' but Object carries`;
+      const result = filterNoMatchReason(binding, kubernetesObject, capabilityNamespaces);
+      expect(result).toEqual(
+        typeof kubernetesObject.metadata === "object" && obj !== null && Object.keys(obj).length > 0
+          ? `${expectedErrorMessage} '${kubernetesObject.metadata.namespace}'.`
+          : `${expectedErrorMessage} ''.`,
+      );
+    },
+  );
+
+  describe("when pod namespace matches the namespace regex", () => {
+    it.each([["pepr-system"], ["pepr-uds-system"], ["uds-system"], ["some-thing-that-is-a-system"], ["your-system"]])(
+      "should not return an error message (namespace: '%s')",
+      namespace => {
+        const binding: Binding = {
+          ...defaultBinding,
+          kind: { kind: "Pod", group: "some-group" },
+          filters: {
+            ...defaultFilters,
+            regexName: "",
+            regexNamespaces: [new RegExp("(.*)-system").source],
+            namespaces: [],
+          },
+        };
+        const kubernetesObject: KubernetesObject = { ...defaultKubernetesObject, metadata: { namespace: namespace } };
+        const capabilityNamespaces: string[] = [];
+        const result = filterNoMatchReason(binding, kubernetesObject, capabilityNamespaces);
+        expect(result).toEqual("");
+      },
     );
   });
-});
 
-it("create: should reject when regex name does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "^default$",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines name regex '.*' but Object carries '.*'./,
-  );
-});
+  // Names Fail
+  it("returns regex name filter error for Pods whos name does not match the regex", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      kind: { kind: "Pod", group: "some-group" },
+      filters: { ...defaultFilters, regexName: "^system", namespaces: [] },
+    };
+    const obj = { metadata: { name: "pepr-demo" } };
+    const objArray = [
+      { ...obj },
+      { ...obj, metadata: { name: "pepr-uds" } },
+      { ...obj, metadata: { name: "pepr-core" } },
+      { ...obj, metadata: { name: "uds-ns" } },
+      { ...obj, metadata: { name: "uds" } },
+    ];
+    const capabilityNamespaces: string[] = [];
+    objArray.map(object => {
+      const result = filterNoMatchReason(binding, object as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+      expect(result).toEqual(
+        `Ignoring Watch Callback: Binding defines name regex '^system' but Object carries '${object?.metadata?.name}'.`,
+      );
+    });
+  });
 
-it("create: should not reject when regex name does match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "^cool",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
+  // Names Pass
+  it("returns no regex name filter error for Pods whos name does match the regex", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      kind: { kind: "Pod", group: "some-group" },
+      filters: { ...defaultFilters, regexName: "^system" },
+    };
+    const obj = { metadata: { name: "pepr-demo" } };
+    const objArray = [
+      { ...obj, metadata: { name: "systemd" } },
+      { ...obj, metadata: { name: "systemic" } },
+      { ...obj, metadata: { name: "system-of-kube-apiserver" } },
+      { ...obj, metadata: { name: "system" } },
+      { ...obj, metadata: { name: "system-uds" } },
+    ];
+    const capabilityNamespaces: string[] = [];
+    objArray.map(object => {
+      const result = filterNoMatchReason(binding, object as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+      expect(result).toEqual(``);
+    });
+  });
 
-it("delete: should reject when regex name does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "^default$",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestDeletePod();
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines name regex '.*' but Object carries '.*'./,
-  );
-});
+  describe("when capability namespaces are present", () => {
+    it("should return missingCarriableNamespace filter error for cluster-scoped objects", () => {
+      const binding: Binding = {
+        ...defaultBinding,
+        filters: { ...defaultFilters, regexName: "" },
+        kind: { kind: "ClusterRole", group: "some-group" },
+      };
+      const obj: KubernetesObject = {
+        kind: "ClusterRole",
+        apiVersion: "rbac.authorization.k8s.io/v1",
+        metadata: { name: "clusterrole1" },
+      };
+      const capabilityNamespaces: string[] = ["monitoring"];
+      const result = filterNoMatchReason(binding, obj, capabilityNamespaces);
+      expect(result).toEqual(
+        "Ignoring Watch Callback: Object does not carry a namespace but namespaces allowed by Capability are '[\"monitoring\"]'.",
+      );
+    });
+  });
 
-it("delete: should not reject when regex name does match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "^cool",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestDeletePod();
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
+  it("returns mismatchedNamespace filter error for clusterScoped objects with namespace filters", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      kind: { kind: "ClusterRole", group: "some-group" },
+      filters: { ...defaultFilters, namespaces: ["ns1"] },
+    };
+    const obj = {
+      kind: "ClusterRole",
+      apiVersion: "rbac.authorization.k8s.io/v1",
+      metadata: { name: "clusterrole1" },
+    };
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual("Ignoring Watch Callback: Binding defines namespaces '[\"ns1\"]' but Object carries ''.");
+  });
 
-it("create: should not reject when regex namespace does match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [new RegExp("^helm").source],
-      regexName: "",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
+  it("returns namespace filter error for namespace objects with namespace filters", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      kind: { kind: "Namespace", group: "some-group" },
+      filters: { ...defaultFilters, namespaces: ["ns1"] },
+    };
+    const obj = {};
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual("Ignoring Watch Callback: Cannot use namespace filter on a namespace object.");
+  });
 
-it("create: should reject when regex namespace does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [new RegExp("^argo").source],
-      regexName: "",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines namespace regexes '.*' but Object carries '.*'./,
-  );
-});
-
-it("delete: should reject when regex namespace does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [new RegExp("^argo").source],
-      regexName: "",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestDeletePod();
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines namespace regexes '.*' but Object carries '.*'./,
-  );
-});
-
-it("delete: should not reject when regex namespace does match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [new RegExp("^helm").source],
-      regexName: "",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestDeletePod();
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
-
-it("delete: should reject when name does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "bleh",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "^not-cool",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestDeletePod();
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines name '.*' but Object carries '.*'./,
-  );
-});
-
-it("should reject when kind does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: {
-      group: "",
-      version: "v1",
-      kind: "Nope",
-    },
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines kind '.*' but Request declares '.*'./,
-  );
-});
-
-it("should reject when group does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: {
-      group: "Nope",
-      version: "v1",
-      kind: "Pod",
-    },
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines group '.*' but Request declares '.*'./,
-  );
-});
-
-it("should reject when version does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: {
-      group: "",
-      version: "Nope",
-      kind: "Pod",
-    },
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines version '.*' but Request declares '.*'./,
-  );
-});
-
-it("should allow when group, version, and kind match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
-
-it("should allow when kind match and others are empty", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: {
-      group: "",
-      version: "",
-      kind: "Pod",
-    },
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
-
-it("should reject when the capability namespace does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, ["bleh", "bleh2"])).toMatch(
-    /Ignoring Admission Callback: Object carries namespace '.*' but namespaces allowed by Capability are '.*'./,
-  );
-});
-
-it("should reject when namespace does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: ["bleh"],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines namespaces '.*' but Object carries '.*'./,
-  );
-});
-
-it("should allow when namespace is match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: ["helm-releasename", "unicorn", "things"],
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
-
-it("should reject when label does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {
-        foo: "bar",
+  it("return an Ignoring Watch Callback string if the binding name and object name are different", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, name: "pepr" },
+    };
+    const obj = {
+      metadata: {
+        name: "not-pepr",
       },
-      annotations: {},
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
+    };
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual(`Ignoring Watch Callback: Binding defines name 'pepr' but Object carries 'not-pepr'.`);
+  });
 
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines labels '.*' but Object carries '.*'./,
-  );
-});
+  describe("when the binding name and KubernetesObject name are the same", () => {
+    it("should not return an Ignoring Watch Callback message", () => {
+      const binding: Binding = {
+        ...defaultBinding,
+        filters: { ...defaultFilters, regexName: "", name: "pepr" },
+      };
+      const obj: KubernetesObject = {
+        metadata: { name: "pepr" },
+      };
+      const capabilityNamespaces: string[] = [];
+      const result = filterNoMatchReason(binding, obj, capabilityNamespaces);
+      expect(result).toEqual("");
+    });
+  });
 
-it("should allow when label is match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      deletionTimestamp: false,
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "",
-      labels: {
-        foo: "bar",
-        test: "test1",
+  it("return deletionTimestamp error when there is no deletionTimestamp in the object", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, deletionTimestamp: true },
+    };
+    const obj = {
+      metadata: {},
+    };
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual("Ignoring Watch Callback: Binding defines deletionTimestamp but Object does not carry it.");
+  });
+
+  it("return no deletionTimestamp error when there is a deletionTimestamp in the object", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, deletionTimestamp: true },
+    };
+    const obj = {
+      metadata: {
+        deletionTimestamp: "2021-01-01T00:00:00Z",
       },
-      annotations: {},
-    },
-    callback,
-  };
+    };
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).not.toEqual("Ignoring Watch Callback: Binding defines deletionTimestamp Object does not carry it.");
+  });
 
-  const pod = AdmissionRequestCreatePod();
-  pod.object.metadata = pod.object.metadata || {};
-  pod.object.metadata.labels = {
-    foo: "bar",
-    test: "test1",
-    test2: "test2",
-  };
+  it("returns label overlap error when there is no overlap between binding and object labels", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, labels: { key: "value" } },
+    };
+    const obj = {
+      metadata: { labels: { anotherKey: "anotherValue" } },
+    };
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual(
+      `Ignoring Watch Callback: Binding defines labels '{"key":"value"}' but Object carries '{"anotherKey":"anotherValue"}'.`,
+    );
+  });
 
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
+  it("returns annotation overlap error when there is no overlap between binding and object annotations", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, annotations: { key: "value" } },
+    };
+    const obj = {
+      metadata: { annotations: { anotherKey: "anotherValue" } },
+    };
+    const capabilityNamespaces: string[] = [];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual(
+      `Ignoring Watch Callback: Binding defines annotations '{"key":"value"}' but Object carries '{"anotherKey":"anotherValue"}'.`,
+    );
+  });
 
-it("should reject when annotation does not match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {
-        foo: "bar",
+  it("returns capability namespace error when object is not in capability namespaces", () => {
+    const binding: Binding = {
+      model: kind.Pod,
+      event: Event.ANY,
+      kind: {
+        group: "",
+        version: "v1",
+        kind: "Pod",
       },
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
-  const pod = AdmissionRequestCreatePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines annotations '.*' but Object carries '.*'./,
-  );
-});
-
-it("should allow when annotation is match", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      annotations: {
-        foo: "bar",
-        test: "test1",
+      filters: {
+        name: "bleh",
+        namespaces: [],
+        regexNamespaces: [],
+        regexName: "",
+        labels: {},
+        annotations: {},
+        deletionTimestamp: false,
       },
-      deletionTimestamp: false,
-      regexNamespaces: [],
-      regexName: "",
-    },
-    callback,
-  };
+      watchCallback: callback,
+    };
 
-  const pod = AdmissionRequestCreatePod();
-  pod.object.metadata = pod.object.metadata || {};
-  pod.object.metadata.annotations = {
-    foo: "bar",
-    test: "test1",
-    test2: "test2",
-  };
+    const obj = {
+      metadata: { namespace: "ns2", name: "bleh" },
+    };
+    const capabilityNamespaces = ["ns1"];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual(
+      `Ignoring Watch Callback: Object carries namespace 'ns2' but namespaces allowed by Capability are '["ns1"]'.`,
+    );
+  });
 
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
+  it("returns binding namespace error when filter namespace is not part of capability namespaces", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, namespaces: ["ns3"], regexNamespaces: [] },
+    };
+    const obj = {};
+    const capabilityNamespaces = ["ns1", "ns2"];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual(
+      `Ignoring Watch Callback: Binding defines namespaces ["ns3"] but namespaces allowed by Capability are '["ns1","ns2"]'.`,
+    );
+  });
 
-it("should use `oldObject` when the operation is `DELETE`", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.DELETE,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "",
-      deletionTimestamp: false,
-      labels: {
-        "test-op": "delete",
+  it("returns binding and object namespace error when they do not overlap", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, namespaces: ["ns1"], regexNamespaces: [] },
+    };
+    const obj = {
+      metadata: { namespace: "ns2" },
+    };
+    const capabilityNamespaces = ["ns1", "ns2"];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual(`Ignoring Watch Callback: Binding defines namespaces '["ns1"]' but Object carries 'ns2'.`);
+  });
+
+  describe("when a KubernetesObject is in an ingnored namespace", () => {
+    it("should return a watch violation message", () => {
+      const binding: Binding = {
+        ...defaultBinding,
+        filters: { ...defaultFilters, regexName: "", namespaces: ["ns3"] },
+      };
+      const kubernetesObject: KubernetesObject = {
+        ...defaultKubernetesObject,
+        metadata: { namespace: "ns3" },
+      };
+      const capabilityNamespaces = ["ns3"];
+      const ignoredNamespaces = ["ns3"];
+      const result = filterNoMatchReason(binding, kubernetesObject, capabilityNamespaces, ignoredNamespaces);
+      expect(result).toEqual(
+        `Ignoring Watch Callback: Object carries namespace 'ns3' but ignored namespaces include '["ns3"]'.`,
+      );
+    });
+  });
+
+  it("returns empty string when all checks pass", () => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: {
+        ...defaultFilters,
+        regexName: "",
+        namespaces: ["ns1"],
+        labels: { key: "value" },
+        annotations: { key: "value" },
       },
-      annotations: {},
-    },
-    callback,
-  };
-
-  const pod = AdmissionRequestDeletePod();
-
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
-
-it("should allow when deletionTimestamp is present on pod", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      regexNamespaces: [],
-      regexName: "",
-      annotations: {
-        foo: "bar",
-        test: "test1",
-      },
-      deletionTimestamp: true,
-    },
-    callback,
-  };
-
-  const pod = AdmissionRequestCreatePod();
-  pod.object.metadata = pod.object.metadata || {};
-  pod.object.metadata!.deletionTimestamp = new Date("2021-09-01T00:00:00Z");
-  pod.object.metadata.annotations = {
-    foo: "bar",
-    test: "test1",
-    test2: "test2",
-  };
-
-  expect(shouldSkipRequest(binding, pod, [])).toBe("");
-});
-
-it("should reject when deletionTimestamp is not present on pod", () => {
-  const binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: podKind,
-    filters: {
-      name: "",
-      namespaces: [],
-      labels: {},
-      regexNamespaces: [],
-      regexName: "",
-      annotations: {
-        foo: "bar",
-        test: "test1",
-      },
-      deletionTimestamp: true,
-    },
-    callback,
-  };
-
-  const pod = AdmissionRequestCreatePod();
-  pod.object.metadata = pod.object.metadata || {};
-  pod.object.metadata.annotations = {
-    foo: "bar",
-    test: "test1",
-    test2: "test2",
-  };
-
-  expect(shouldSkipRequest(binding, pod, [])).toMatch(
-    /Ignoring Admission Callback: Binding defines deletionTimestamp but Object does not carry it./,
-  );
+    };
+    const obj = {
+      metadata: { namespace: "ns1", labels: { key: "value" }, annotations: { key: "value" } },
+    };
+    const capabilityNamespaces = ["ns1"];
+    const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
+    expect(result).toEqual("");
+  });
 });
 
 describe("adjudicateMisboundDeleteWithDeletionTimestamp", () => {
@@ -993,5 +1352,37 @@ describe("adjudicateMismatchedNamespaceRegex", () => {
       { metadata: { namespace: "default" } },
     );
     expect(result).toBe(null);
+  });
+});
+
+describe("adjudicateMisboundNamespace", () => {
+  const defaultBinding: Binding = {
+    event: Event.CREATE,
+    filters: {
+      annotations: {},
+      deletionTimestamp: false,
+      labels: {},
+      name: "",
+      namespaces: [],
+      regexName: "^default$",
+      regexNamespaces: [],
+    },
+    kind: {
+      group: "v1",
+      kind: "Namespace",
+    },
+    model: kind.Namespace,
+  };
+
+  it("should return nothing when binding is correct", () => {
+    const result = adjudicateMisboundNamespace(defaultBinding);
+    expect(result).toBe(null);
+  });
+
+  it("should return reason when binding is incorrect", () => {
+    const testBinding = clone(defaultBinding);
+    testBinding.filters.namespaces = ["oof"];
+    const result = adjudicateMisboundNamespace(testBinding);
+    expect(result).toBe(`Cannot use namespace filter on a namespace object.`);
   });
 });
