+++ src/lib/helpers.test.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { Binding, CapabilityExport } from "./types";
+import { CapabilityExport } from "./types";
 import { Event } from "./enums";
 import {
   bindingAndCapabilityNSConflict,
@@ -20,33 +20,13 @@ import {
   validateCapabilityNames,
   ValidationError,
 } from "./helpers";
-import { filterNoMatchReason } from "./filter/filterNoMatchReason";
 import { sanitizeResourceName } from "../sdk/sdk";
 import * as fc from "fast-check";
 import { expect, describe, jest, beforeEach, afterEach, it } from "@jest/globals";
 import { SpiedFunction } from "jest-mock";
-import { KubernetesObject, kind, modelToGroupVersionKind } from "kubernetes-fluent-client";
-import { defaultFilters, defaultKubernetesObject, defaultBinding } from "./filter/adjudicators/defaultTestObjects";
+import { kind } from "kubernetes-fluent-client";
 
 export const callback = () => undefined;
-
-export const podKind = modelToGroupVersionKind(kind.Pod.name);
-export const deploymentKind = modelToGroupVersionKind(kind.Deployment.name);
-export const clusterRoleKind = modelToGroupVersionKind(kind.ClusterRole.name);
-
-export const groupBinding: Binding = {
-  event: Event.CREATE,
-  filters: defaultFilters,
-  kind: deploymentKind,
-  model: kind.Deployment,
-};
-
-export const clusterScopedBinding: Binding = {
-  event: Event.DELETE,
-  filters: defaultFilters,
-  kind: clusterRoleKind,
-  model: kind.ClusterRole,
-};
 
 const mockCapabilities: CapabilityExport[] = JSON.parse(`[
     {
@@ -778,339 +758,6 @@ describe("replaceString", () => {
     const expected = "Remove.";
     expect(replaceString(original, stringA, stringB)).toBe(expected);
   });
-});
-
-describe("filterNoMatchReason", () => {
-  it.each([
-    [{}],
-    [{ metadata: { namespace: "pepr-uds" } }],
-    [{ metadata: { namespace: "pepr-core" } }],
-    [{ metadata: { namespace: "uds-ns" } }],
-    [{ metadata: { namespace: "uds" } }],
-  ])(
-    "given %j, it returns regex namespace filter error for Pods whose namespace does not match the regex",
-    (obj: KubernetesObject) => {
-      const kubernetesObject: KubernetesObject = obj.metadata
-        ? {
-            ...defaultKubernetesObject,
-            metadata: { ...defaultKubernetesObject.metadata, namespace: obj.metadata.namespace },
-          }
-        : { ...defaultKubernetesObject, metadata: obj as unknown as undefined };
-      const binding: Binding = {
-        ...defaultBinding,
-        kind: { kind: "Pod", group: "some-group" },
-        filters: { ...defaultFilters, regexNamespaces: [new RegExp("(.*)-system").source] },
-      };
-
-      const capabilityNamespaces: string[] = [];
-      const expectedErrorMessage = `Ignoring Watch Callback: Binding defines namespace regexes '["(.*)-system"]' but Object carries`;
-      const result = filterNoMatchReason(binding, kubernetesObject, capabilityNamespaces);
-      expect(result).toEqual(
-        typeof kubernetesObject.metadata === "object" && obj !== null && Object.keys(obj).length > 0
-          ? `${expectedErrorMessage} '${kubernetesObject.metadata.namespace}'.`
-          : `${expectedErrorMessage} ''.`,
-      );
-    },
-  );
-});
-
-describe("when pod namespace matches the namespace regex", () => {
-  it.each([["pepr-system"], ["pepr-uds-system"], ["uds-system"], ["some-thing-that-is-a-system"], ["your-system"]])(
-    "should not return an error message (namespace: '%s')",
-    namespace => {
-      const binding: Binding = {
-        ...defaultBinding,
-        kind: { kind: "Pod", group: "some-group" },
-        filters: {
-          ...defaultFilters,
-          regexName: "",
-          regexNamespaces: [new RegExp("(.*)-system").source],
-          namespaces: [],
-        },
-      };
-      const kubernetesObject: KubernetesObject = { ...defaultKubernetesObject, metadata: { namespace: namespace } };
-      const capabilityNamespaces: string[] = [];
-      const result = filterNoMatchReason(binding, kubernetesObject, capabilityNamespaces);
-      expect(result).toEqual("");
-    },
-  );
-});
-
-// Names Fail
-it("returns regex name filter error for Pods whos name does not match the regex", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    kind: { kind: "Pod", group: "some-group" },
-    filters: { ...defaultFilters, regexName: "^system", namespaces: [] },
-  };
-  const obj = { metadata: { name: "pepr-demo" } };
-  const objArray = [
-    { ...obj },
-    { ...obj, metadata: { name: "pepr-uds" } },
-    { ...obj, metadata: { name: "pepr-core" } },
-    { ...obj, metadata: { name: "uds-ns" } },
-    { ...obj, metadata: { name: "uds" } },
-  ];
-  const capabilityNamespaces: string[] = [];
-  objArray.map(object => {
-    const result = filterNoMatchReason(binding, object as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Binding defines name regex '^system' but Object carries '${object?.metadata?.name}'.`,
-    );
-  });
-});
-
-// Names Pass
-it("returns no regex name filter error for Pods whos name does match the regex", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    kind: { kind: "Pod", group: "some-group" },
-    filters: { ...defaultFilters, regexName: "^system" },
-  };
-  const obj = { metadata: { name: "pepr-demo" } };
-  const objArray = [
-    { ...obj, metadata: { name: "systemd" } },
-    { ...obj, metadata: { name: "systemic" } },
-    { ...obj, metadata: { name: "system-of-kube-apiserver" } },
-    { ...obj, metadata: { name: "system" } },
-    { ...obj, metadata: { name: "system-uds" } },
-  ];
-  const capabilityNamespaces: string[] = [];
-  objArray.map(object => {
-    const result = filterNoMatchReason(binding, object as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-    expect(result).toEqual(``);
-  });
-});
-
-describe("when capability namespaces are present", () => {
-  it("should return missingCarriableNamespace filter error for cluster-scoped objects", () => {
-    const binding: Binding = {
-      ...defaultBinding,
-      filters: { ...defaultFilters, regexName: "" },
-      kind: { kind: "ClusterRole", group: "some-group" },
-    };
-    const obj: KubernetesObject = {
-      kind: "ClusterRole",
-      apiVersion: "rbac.authorization.k8s.io/v1",
-      metadata: { name: "clusterrole1" },
-    };
-    const capabilityNamespaces: string[] = ["monitoring"];
-    const result = filterNoMatchReason(binding, obj, capabilityNamespaces);
-    expect(result).toEqual(
-      "Ignoring Watch Callback: Object does not carry a namespace but namespaces allowed by Capability are '[\"monitoring\"]'.",
-    );
-  });
-});
-
-it("returns mismatchedNamespace filter error for clusterScoped objects with namespace filters", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    kind: { kind: "ClusterRole", group: "some-group" },
-    filters: { ...defaultFilters, namespaces: ["ns1"] },
-  };
-  const obj = {
-    kind: "ClusterRole",
-    apiVersion: "rbac.authorization.k8s.io/v1",
-    metadata: { name: "clusterrole1" },
-  };
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual("Ignoring Watch Callback: Binding defines namespaces '[\"ns1\"]' but Object carries ''.");
-});
-
-it("returns namespace filter error for namespace objects with namespace filters", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    kind: { kind: "Namespace", group: "some-group" },
-    filters: { ...defaultFilters, namespaces: ["ns1"] },
-  };
-  const obj = {};
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual("Ignoring Watch Callback: Cannot use namespace filter on a namespace object.");
-});
-
-it("return an Ignoring Watch Callback string if the binding name and object name are different", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, name: "pepr" },
-  };
-  const obj = {
-    metadata: {
-      name: "not-pepr",
-    },
-  };
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual(`Ignoring Watch Callback: Binding defines name 'pepr' but Object carries 'not-pepr'.`);
-});
-
-describe("when the binding name and KubernetesObject name are the same", () => {
-  it("should not return an Ignoring Watch Callback message", () => {
-    const binding: Binding = {
-      ...defaultBinding,
-      filters: { ...defaultFilters, regexName: "", name: "pepr" },
-    };
-    const obj: KubernetesObject = {
-      metadata: { name: "pepr" },
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(binding, obj, capabilityNamespaces);
-    expect(result).toEqual("");
-  });
-});
-
-it("return deletionTimestamp error when there is no deletionTimestamp in the object", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, deletionTimestamp: true },
-  };
-  const obj = {
-    metadata: {},
-  };
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual("Ignoring Watch Callback: Binding defines deletionTimestamp but Object does not carry it.");
-});
-
-it("return no deletionTimestamp error when there is a deletionTimestamp in the object", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, deletionTimestamp: true },
-  };
-  const obj = {
-    metadata: {
-      deletionTimestamp: "2021-01-01T00:00:00Z",
-    },
-  };
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).not.toEqual("Ignoring Watch Callback: Binding defines deletionTimestamp Object does not carry it.");
-});
-
-it("returns label overlap error when there is no overlap between binding and object labels", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, labels: { key: "value" } },
-  };
-  const obj = {
-    metadata: { labels: { anotherKey: "anotherValue" } },
-  };
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual(
-    `Ignoring Watch Callback: Binding defines labels '{"key":"value"}' but Object carries '{"anotherKey":"anotherValue"}'.`,
-  );
-});
-
-it("returns annotation overlap error when there is no overlap between binding and object annotations", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, annotations: { key: "value" } },
-  };
-  const obj = {
-    metadata: { annotations: { anotherKey: "anotherValue" } },
-  };
-  const capabilityNamespaces: string[] = [];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual(
-    `Ignoring Watch Callback: Binding defines annotations '{"key":"value"}' but Object carries '{"anotherKey":"anotherValue"}'.`,
-  );
-});
-
-it("returns capability namespace error when object is not in capability namespaces", () => {
-  const binding: Binding = {
-    model: kind.Pod,
-    event: Event.ANY,
-    kind: {
-      group: "",
-      version: "v1",
-      kind: "Pod",
-    },
-    filters: {
-      name: "bleh",
-      namespaces: [],
-      regexNamespaces: [],
-      regexName: "",
-      labels: {},
-      annotations: {},
-      deletionTimestamp: false,
-    },
-    watchCallback: callback,
-  };
-
-  const obj = {
-    metadata: { namespace: "ns2", name: "bleh" },
-  };
-  const capabilityNamespaces = ["ns1"];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual(
-    `Ignoring Watch Callback: Object carries namespace 'ns2' but namespaces allowed by Capability are '["ns1"]'.`,
-  );
-});
-
-it("returns binding namespace error when filter namespace is not part of capability namespaces", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, namespaces: ["ns3"], regexNamespaces: [] },
-  };
-  const obj = {};
-  const capabilityNamespaces = ["ns1", "ns2"];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual(
-    `Ignoring Watch Callback: Binding defines namespaces ["ns3"] but namespaces allowed by Capability are '["ns1","ns2"]'.`,
-  );
-});
-
-it("returns binding and object namespace error when they do not overlap", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: { ...defaultFilters, namespaces: ["ns1"], regexNamespaces: [] },
-  };
-  const obj = {
-    metadata: { namespace: "ns2" },
-  };
-  const capabilityNamespaces = ["ns1", "ns2"];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual(`Ignoring Watch Callback: Binding defines namespaces '["ns1"]' but Object carries 'ns2'.`);
-});
-
-describe("when a KubernetesObject is in an ingnored namespace", () => {
-  it("should return a watch violation message", () => {
-    const binding: Binding = {
-      ...defaultBinding,
-      filters: { ...defaultFilters, regexName: "", namespaces: ["ns3"] },
-    };
-    const kubernetesObject: KubernetesObject = {
-      ...defaultKubernetesObject,
-      metadata: { namespace: "ns3" },
-    };
-    const capabilityNamespaces = ["ns3"];
-    const ignoredNamespaces = ["ns3"];
-    const result = filterNoMatchReason(binding, kubernetesObject, capabilityNamespaces, ignoredNamespaces);
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Object carries namespace 'ns3' but ignored namespaces include '["ns3"]'.`,
-    );
-  });
-});
-
-it("returns empty string when all checks pass", () => {
-  const binding: Binding = {
-    ...defaultBinding,
-    filters: {
-      ...defaultFilters,
-      regexName: "",
-      namespaces: ["ns1"],
-      labels: { key: "value" },
-      annotations: { key: "value" },
-    },
-  };
-  const obj = {
-    metadata: { namespace: "ns1", labels: { key: "value" }, annotations: { key: "value" } },
-  };
-  const capabilityNamespaces = ["ns1"];
-  const result = filterNoMatchReason(binding, obj as unknown as Partial<KubernetesObject>, capabilityNamespaces);
-  expect(result).toEqual("");
 });
 
 describe("validateHash", () => {
